refactor: change export hook so it can be used in audit and in general export
